Remove unncessary span which breaks titles for me.
At least in my version of Chrome (51.0.2704.106) having a title on a span
nested within a div does not result in the title showing upon hover. However
putting the title on the div works. Given that I can't see any purpose for the
nested span, moving the title to the div and removing the span is a quick fix.